helix/requirements.txt
openai-whisper
openwakeword
sounddevice
numpy
onnxruntime
pynetworktables
